+++ src/Model/Order.php
@@ -157,9 +157,10 @@ class Order extends Order_parent
                 $remark = 'AmazonPay: ERROR';
                 if (!empty($data['result']['response'])) {
                     $response = PhpHelper::jsonToArray($data['result']['response']);
-                    $remark = 'AmazonPay ERROR: ' . $response['reasonCode'];
+                    $remark .= ' (' . $response['reasonCode'] . ')';
                 }
-
+                $this->_setFieldData('oxtransstatus', 'NOT_FINISHED');
+                $this->_setFieldData('oxfolder', 'ORDERFOLDER_PROBLEMS');
                 $this->_setFieldData('osc_amazon_remark', $remark);
                 $this->save();
                 break;
@@ -187,6 +188,19 @@ class Order extends Order_parent
                 $this->_setFieldData('oxfolder', 'ORDERFOLDER_NEW');
                 $this->save();
                 break;
+
+            case "AMZ_AUTH_OR_CAPT_DECLINED":
+                $remark = 'AmazonPay: Auth or Capture Declined';
+                if (!empty($data['result']['response'])) {
+                    $response = PhpHelper::jsonToArray($data['result']['response']);
+                    $remark .= ' (' . $response['reasonCode'] . ')';
+                }
+                $this->_setFieldData('oxtransstatus', 'NOT_FINISHED');
+                $this->_setFieldData('oxfolder', 'ORDERFOLDER_PROBLEMS');
+                $this->_setFieldData('osc_amazon_remark', $remark);
+                $this->save();
+                break;
+
         }
     }